refactor(InputValidator): 메서드 접근제어자 수정 public->private
